+++ src/repositories/types/license.rs
@@ -1,6 +1,59 @@
+use regex::Regex;
 // SPDX-License-Identifier: GPL-3.0-only
-use serde::{Deserialize, Serialize};
-use std::fmt::Display;
+use serde::{Deserialize, Serialize, de};
+use std::{fmt::Display, str::FromStr, sync::LazyLock};
+use thiserror::Error;
+
+const VALID_LICENSE: &str = r"^[a-zA-Z0-9\-\+\.]+$";
+static LICENSE_REGEX: LazyLock<Regex> = LazyLock::new(|| Regex::new(VALID_LICENSE).expect("Expected valid regex"));
+
+/// Errors that occur when creating or parsing the license identifier.
+#[cfg_attr(test, derive(PartialEq))]
+#[derive(Error, Debug)]
+pub enum LicenseError {
+    #[error("Invalid license identifier '{0}', cannot be empty and can only contain characters: 'a-z', 'A-Z', '0-9', '-', '+' and '.'")]
+    InvalidLicense(String),
+}
+
+/// Represents a license identifier.
+#[derive(Serialize, Debug, Clone)]
+pub struct LicenseIdentifier(String);
+
+impl<'de> Deserialize<'de> for LicenseIdentifier {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        let string: String = de::Deserialize::deserialize(deserializer)?;
+
+        Self::from_str(&string).map_err(de::Error::custom)
+    }
+}
+
+impl FromStr for LicenseIdentifier {
+    type Err = LicenseError;
+
+    fn from_str(string: &str) -> Result<Self, Self::Err> {
+        if !LICENSE_REGEX.is_match(string) {
+            return Err(LicenseError::InvalidLicense(string.to_string()));
+        }
+
+        Ok(Self(string.to_string()))
+    }
+}
+
+impl Display for LicenseIdentifier {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.0)?;
+        Ok(())
+    }
+}
+
+impl PartialEq for LicenseIdentifier {
+    fn eq(&self, other: &Self) -> bool {
+        self.0.to_lowercase() == other.0.to_lowercase()
+    }
+}
 
 /// Wrapper to differentiate between different License types in metadata files.
 #[cfg_attr(test, derive(PartialEq))]
@@ -9,9 +62,9 @@ use std::fmt::Display;
 pub enum Licenses {
     #[default]
     Unknown,
-    Single(String),
+    Single(LicenseIdentifier),
     SingleWithExceptions {
-        name: String,
+        name: LicenseIdentifier,
 
         #[serde(rename = "with")]
         exceptions: Vec<String>,
@@ -97,26 +150,51 @@ pub mod tests {
     use super::*;
 
     #[test]
+    fn valid_from_str() {
+        let identifier = &"Test-1.1+";
+        assert_eq!(
+            LicenseIdentifier::from_str(identifier),
+            Ok(LicenseIdentifier(identifier.to_string()))
+        );
+    }
+
+    #[test]
+    fn from_str_no_input() {
+        assert_eq!(LicenseIdentifier::from_str(""), Err(LicenseError::InvalidLicense("".to_string())));
+    }
+
+    #[test]
+    fn from_str_illegal_chars() {
+        let illegal_chars = " _/\\!@#$%^&*():;'\"<>[]{}?|~`±§=\u{1234}";
+        for identifier in illegal_chars.chars() {
+            assert_eq!(
+                LicenseIdentifier::from_str(&identifier.to_string()),
+                Err(LicenseError::InvalidLicense(identifier.to_string())),
+            );
+        }
+    }
+
+    #[test]
     fn unknown_check() {
         let license = Licenses::Unknown;
         assert!(license.is_unknown());
 
-        let license = Licenses::Single("MIT".to_string());
+        let license = Licenses::Single(LicenseIdentifier("MIT".into()));
         assert!(!license.is_unknown());
     }
 
     #[test]
     fn length() {
         assert_eq!(Licenses::Unknown.get_length(), 0);
-        assert_eq!(Licenses::Single("MIT".to_string()).get_length(), 1);
+        assert_eq!(Licenses::Single(LicenseIdentifier("MIT".into())).get_length(), 1);
 
         let license = Licenses::SingleWithExceptions {
-            name: "MIT".to_string(),
+            name: LicenseIdentifier("MIT".into()),
             exceptions: vec!["Test-Exception".to_string()],
         };
         assert_eq!(license.get_length(), 1);
 
-        let license = Licenses::Single("Test".to_string());
+        let license = Licenses::Single(LicenseIdentifier("Test".into()));
         let recursive_license = Licenses::All {
             all: vec![license.clone(), license.clone()],
         };
@@ -137,14 +215,14 @@ pub mod tests {
 
     #[test]
     fn format_single() {
-        let license = Licenses::Single("MIT".to_string());
+        let license = Licenses::Single(LicenseIdentifier("MIT".into()));
         assert_eq!(license.to_string(), "MIT".to_string());
     }
 
     #[test]
     fn format_single_with_single_exception() {
         let license = Licenses::SingleWithExceptions {
-            name: "MIT".to_string(),
+            name: LicenseIdentifier("MIT".into()),
             exceptions: vec!["Test-Exception".to_string()],
         };
         assert_eq!(license.to_string(), "MIT WITH Test-Exception".to_string());
@@ -153,7 +231,7 @@ pub mod tests {
     #[test]
     fn format_single_with_exceptions() {
         let license = Licenses::SingleWithExceptions {
-            name: "MIT".to_string(),
+            name: LicenseIdentifier("MIT".into()),
             exceptions: vec!["Test-Exception".to_string(), "Second-Exception".to_string()],
         };
         assert_eq!(license.to_string(), "MIT WITH (Test-Exception, Second-Exception)".to_string());
@@ -162,7 +240,7 @@ pub mod tests {
     #[test]
     fn format_single_with_empty_exceptions() {
         let license = Licenses::SingleWithExceptions {
-            name: "MIT".to_string(),
+            name: LicenseIdentifier("MIT".into()),
             exceptions: Vec::new(),
         };
         assert_eq!(license.to_string(), "MIT".to_string());
@@ -170,7 +248,7 @@ pub mod tests {
 
     #[test]
     fn format_recursive() {
-        let license = Licenses::Single("Test".to_string());
+        let license = Licenses::Single(LicenseIdentifier("Test".into()));
         let recursive_license = Licenses::All {
             all: vec![license.clone()],
         };